requirements.txt
aiofiles==23.1.0
aiogram==3.0.0rc1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
Django==4.1
djangorestframework==3.14.0
drf-extra-fields==3.7.0
django_ckeditor==6.7.0
filetype==1.2.0
flake8==6.1.0
frozenlist==1.4.0
google-api-python-client==2.97.0
gunicorn==20.0.4
idna==3.4
Jinja2==3.1.2
magic-filter==1.0.11
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
numpy==1.25.2
Pillow==10.0.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
soundfile==0.12.1
SpeechRecognition==3.10.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
bs4==0.0.1
lxml==4.9.3
google==3.0.0
django-admin-rangefilter==0.11.0
csp==0.1a10
bleach